+++ NEWS.md
@@ -9,6 +9,7 @@ TODO
 - Internal refactoring: Use one CSS file for all styles, and use CSS files to load custom CSS of each style (#37)
 - Added a sample Shiny app "demo" that lets you experiment with all loader types and parameters
 - File cleanup: Remove all files and dirs from inst/css-loaders and only keep the CSS (#38)
+- Removed the default value of `proxy.height` parameter to simplify the function documentation
 
 # shinycssloaders 0.3